float deserializer on setPowerMeterInput and setEnergyMeterInput

+++ src/ArduinoOcpp.cpp
@@ -324,8 +324,8 @@ void setEnergyMeterInput(std::function<float()> energyInput, unsigned int connec
     SampledValueProperties meterProperties;
     meterProperties.setMeasurand("Energy.Active.Import.Register");
     meterProperties.setUnit("Wh");
-    auto mvs = std::unique_ptr<SampledValueSamplerConcrete<int32_t, SampledValueDeSerializer<int32_t>>>(
-                           new SampledValueSamplerConcrete<int32_t, SampledValueDeSerializer<int32_t>>(
+    auto mvs = std::unique_ptr<SampledValueSamplerConcrete<float, SampledValueDeSerializer<float>>>(
+                           new SampledValueSamplerConcrete<float, SampledValueDeSerializer<float>>(
             meterProperties,
             [energyInput] (ReadingContext) {return energyInput();}
     ));
@@ -347,8 +347,8 @@ void setPowerMeterInput(std::function<float()> powerInput, unsigned int connecto
     SampledValueProperties meterProperties;
     meterProperties.setMeasurand("Power.Active.Import");
     meterProperties.setUnit("W");
-    auto mvs = std::unique_ptr<SampledValueSamplerConcrete<int32_t, SampledValueDeSerializer<int32_t>>>(
-                           new SampledValueSamplerConcrete<int32_t, SampledValueDeSerializer<int32_t>>(
+    auto mvs = std::unique_ptr<SampledValueSamplerConcrete<float, SampledValueDeSerializer<float>>>(
+                           new SampledValueSamplerConcrete<float, SampledValueDeSerializer<float>>(
             meterProperties,
             [powerInput] (ReadingContext) {return powerInput();}
     ));
